Refactor the private route to use Spotify auth
