+++ app/lib/types/admin.ts
@@ -1,0 +1,4 @@
+export enum AdminFormActions {
+  AddLabel = 'add-label',
+  AddArtistGrouping = 'add-artist-grouping',
+}
