+++ styles/app.scss
@@ -2,6 +2,7 @@
 @tailwind components;
 @tailwind utilities;
 
+a,
 .btn,
 .label {
   cursor: pointer;
